Auth handled in a single function
