+++ src/components/SignIn.jsx
@@ -12,8 +12,9 @@ import Grid from '@mui/material/Grid';
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import Typography from '@mui/material/Typography';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
+
 import SignUp from './SignUp';
-import { authLogin } from './services/authLogin';
+import { authLogin, isLogin } from './services/authLogin';
 import { BrowserRouter as Router, Switch, 
   Route, redirect,  useNavigate} from "react-router-dom";
 import Profile from './Profile';
@@ -38,16 +39,11 @@ export default function SignIn() {
     const email = data.get('email');
     const password = data.get('password');
     const profileObject = {email:email, password:password};
+    
+    isLogin(email, password, profileObject, navigate)
 
-    const auth = authLogin(email).then(function(result) {
-          console.log("result",result)
-          if(result == true){
-            navigate('./Profile', {state: profileObject})
-          }else{
-            window.alert("Invalid Log in");}
-          
-        })
-  };
+    };
+    
 
   return (
     <ThemeProvider theme={theme}>
